+++ .github/workflows/docker/compose/3rd_parties-compose.yaml
@@ -6,3 +6,12 @@ services:
     build:
       dockerfile: comps/3rd_parties/nginx/src/Dockerfile
     image: ${REGISTRY:-opea}/nginx:${TAG:-latest}
+
+  wav2lip:
+    build:
+      dockerfile: comps/3rd_parties/wav2lip/src/Dockerfile
+    image: ${REGISTRY:-opea}/wav2lip:${TAG:-latest}
+  wav2lip-gaudi:
+    build:
+      dockerfile: comps/3rd_parties/wav2lip/src/Dockerfile.intel_hpu
+    image: ${REGISTRY:-opea}/wav2lip-gaudi:${TAG:-latest}
