+++ comps/3rd_parties/wav2lip/src/entrypoint.sh
@@ -23,8 +23,8 @@ export PT_HPU_LAZY_MODE=0
 export PT_HPU_ENABLE_REFINE_DYNAMIC_SHAPES=1
 
 # Wav2Lip, GFPGAN
-cd /home/user/comps/animation/src/integration/ || exit
-python3 dependency/wav2lip_server.py \
+cd /home/user/comps/3rd_parties/wav2lip/src || exit
+python3 wav2lip_server.py \
 --device $DEVICE \
 --port $((WAV2LIP_PORT)) \
 --inference_mode $INFERENCE_MODE \
