--- comps/animation/src/integration/dependency/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
